+++ web/assets/v4.42.0/pwa.js
@@ -1,7 +1,0 @@
-(() => {
-'use strict';
-if (!('serviceWorker' in navigator) || !['http:', 'https:'].includes(location.protocol)) return;
-window.addEventListener('load', () => {
-  navigator.serviceWorker.register('assets/v4.42.0/sw.js', {scope: '/'}).catch(() => {});
-}, {once: true});
-})();